web: rework hostname summary display
The hostname will now show the mode in the Installation Summary section
and the value in the System Information section. This approach avoids
redundant information when both sections are displayed side by side in
the overview.

A prop has been added to allow the hostname value to be displayed on the
confirmation screen, where it would be odd not to show the value. This
is especially relevant since, at the time of writing, the System
Information section is not planned for inclusion on the confirmation
screen.